Isis libraries now using v002
git-svn-id: http://subversion.wr.usgs.gov/repos/prog/isis3/trunk@4072 41f8697f-d340-4b68-9986-7bafba869bb8

+++ isis/src/control/objs/ControlNetVersioner/ControlNetFileV0002.cpp
@@ -196,12 +196,12 @@ namespace Isis {
 
     PvlObject protoCore("Core");
     protoCore.AddKeyword(PvlKeyword("HeaderStartByte",
-        iString(startCoreHeaderPos)));
-    protoCore.AddKeyword(PvlKeyword("HeaderBytes", iString(coreHeaderSize)));
+                         iString((BigInt) startCoreHeaderPos)));
+    protoCore.AddKeyword(PvlKeyword("HeaderBytes", iString((BigInt) coreHeaderSize)));
     protoCore.AddKeyword(PvlKeyword("PointsStartByte",
-        iString(startCoreHeaderPos + coreHeaderSize)));
+        iString((BigInt) ( startCoreHeaderPos + coreHeaderSize))));
     protoCore.AddKeyword(PvlKeyword("PointsBytes",
-        iString(pointsSize)));
+        (BigInt) iString(pointsSize)));
     protoObj.AddObject(protoCore);
 
     PvlGroup netInfo("ControlNetworkInfo");
